% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_bars.R
\name{geom_bars}
\alias{geom_bars}
\title{Bars of Discrete Function using \CRANpkg{ggplot2}}
\usage{
geom_bars(
  fun,
  ...,
  labels = getval_(arg),
  xlim = stop("must provide `xlim` explicitly")
)
}
\arguments{
\item{fun}{a discrete \link[base]{function}}

\item{...}{argument of \link[base]{function} \code{fun} (not of \link[ggplot2]{ggplot})}

\item{labels}{..}

\item{xlim}{\link[base]{length} \eqn{\geq 2} \link[base]{integer} \link[base]{vector}, limit on \eqn{x}-axis.}
}
\value{
Function \link{geom_bars} returns a \link[base]{list} of
\link[ggplot2]{geom_bar}
\link[ggplot2]{layer}s.
}
\description{
..
}
\details{
..
}
\note{
Be aware of potential name clash with \code{ggpubr::ggbarplot}.
}
\examples{
library(ggplot2)
theme_set(theme(
 legend.position = 'inside',
 legend.position.inside = c(.8, .8),
 legend.key.spacing.y = unit(.01, units = 'npc'),
 legend.background = element_rect(color = 'grey95')
))

pdf_ = tempfile(fileext = '.pdf')
cairo_pdf(filename = pdf_)
ggplot() + geom_bars(dpois, lambda = 1, xlim = 0:6)
dev.off()
system(paste('open', pdf_))
file.remove(pdf_)

ggplot() + geom_bars(dnbinom, size = 5.2, prob = c(.5, .4), xlim = 0:15)

ggplot() + geom_bars(dpois, lambda = c(1, 1.3), labels = c('Treatment', 'Control'), 
 xlim = 0:4) +
 labs(x = 'Number of Re-Admissions', y = 'Conditional Probabilities', 
  fill = 'Poisson\nDistribution',
  caption = 'Illustration showing mean-ratio = .77 (Treatment vs. Control)')
 
}
