% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m_atop.R
\name{m_atop}
\alias{m_atop}
\alias{m_atop.expression}
\alias{m_atop.character}
\title{Extension of \code{demo(plotmath)} of Package \pkg{grDevices}}
\usage{
m_atop(x, size, right, accumulate)

\method{m_atop}{expression}(
  x,
  size = c("textstyle", "displaystyle", "scriptstyle", "scriptscriptstyle"),
  right = FALSE,
  accumulate = FALSE
)

\method{m_atop}{character}(
  x,
  size = c("textstyle", "displaystyle", "scriptstyle", "scriptscriptstyle"),
  right = FALSE,
  accumulate = FALSE
)
}
\arguments{
\item{x}{..}

\item{size}{\link[base]{character} scalar, font size, currently supported are
\code{'displaystyle'}, \code{'textstyle'} (default), \code{'scriptstyle'} and \code{'scriptscriptstyle'}}

\item{right, accumulate}{see \link[base]{Reduce}}
}
\value{
Function \link{m_atop} returns a \link[base]{call} to be interpreted as a mathematical expression by
text-drawing functions (e.g., \link[graphics]{title})
}
\description{
..
}
\examples{
txt = c(
'Lower: $log(-\\\\hat{g}(\\\\hat{A} - x_p)/(exp(\\\\hat{g}z_p) - 1))=log(B) + h\\\\cdot z_p^2/2$',
'Upper: $log(\\\\hat{g}(x_{1-p} - \\\\hat{A})/(exp(-\\\\hat{g}z_p) - 1))=log(B) + h\\\\cdot z_p^2/2$',
'Combo: $log(\\\\hat{g}(x_{1-p} - x_p)/(exp(-\\\\hat{g}z_p)-exp(\\\\hat{g}z)))=log(B) + h\\\\cdot z_p^2/2$'
)

library(latex2exp)
(mm <- m_atop(TeX(txt), size = 'scriptstyle'))
stopifnot(identical(mm, m_atop(TeX(txt, output = 'character'), size = 'scriptstyle')))
plot(1, main = mm)

}
\references{
\url{https://stat.ethz.ch/R-manual/R-devel/library/grDevices/html/plotmath.html}.

Inspiration for function \link{m_atop}: \url{https://stackoverflow.com/questions/21067600/multiple-line-breaks-in-expression}
}
\seealso{
\code{demo(plotmath)}

The expressions are processed by \link[graphics]{title}.
}
