% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths_discrete.R
\name{paths_discrete}
\alias{paths_discrete}
\title{Bars of Discrete Function using \CRANpkg{ggplot2}}
\usage{
paths_discrete(
  fun,
  dots = list(),
  args = .mapply(dots, FUN = list, MoreArgs = NULL),
  label,
  xlim = stop("must provide `xlim` explicitly"),
  ...
)
}
\arguments{
\item{fun}{a discrete \link[base]{function}}

\item{dots}{argument of \link[base]{function} \code{fun}}

\item{args}{..}

\item{label}{..}

\item{xlim}{\link[base]{length} \eqn{\geq 2} \link[base]{integer} \link[base]{vector}, limit on \eqn{x}-axis.}

\item{...}{parameters of \link[geomtextpath]{geom_textpath}}
}
\value{
Function \code{\link[=paths_discrete]{paths_discrete()}} returns a \link[base]{list} of
\link[ggplot2]{layer}s.
}
\description{
..
}
\details{
..
}
\note{
Be aware of potential name clash with function \code{ggpubr::ggbarplot}.
}
\examples{
library(ggplot2)
ggplot() + paths_discrete(
 fun = dpois, dots = list(lambda = c(1, 1.3)), 
 label = c('Treatment', 'Control'), xlim = 0:4
) + labs(
 x = 'Number of Re-Admissions', y = 'Conditional Probabilities', 
 caption = 'Illustration showing mean-ratio = .77 (Treatment vs. Control)')
 
}
\keyword{internal}
