% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shareLegend.R
\name{shareLegend}
\alias{shareLegend}
\title{\link[gridExtra]{arrangeGrob} and Share Legend}
\usage{
shareLegend(
  grobs,
  position = c("bottom", "top", "left", "right"),
  top = NULL,
  bottom = NULL,
  left = NULL,
  right = NULL,
  ...
)
}
\arguments{
\item{grobs}{a \link[base]{list} of
\link[grid]{grob}s, \link[gtable]{gtable}s, \link[ggplot2]{ggplot}s or trellis objects}

\item{position}{\link[base]{character} scalar, location of the shared legend.
Currently there is no way to determine whether the \code{grobs} \emph{should} share the same legend.}

\item{top, bottom, left, right}{\link[base]{character} scalars, \link[base]{expression}s or
\link[grid]{grob}s, see details in \link[gridExtra]{arrangeGrob}}

\item{...}{additional parameters of \link[gridExtra]{arrangeGrob},
such as \code{nrow}, \code{ncol}, \code{widths}, \code{heights}}
}
\value{
Function \link{shareLegend} returns a \code{'gtable'} object (returned value of \link[gridExtra]{arrangeGrob})
}
\description{
..
}
\examples{
library(ggplot2)
(p1 = ggplot(data=iris) + geom_point(aes(x=Sepal.Length, y=Petal.Length, color=Species)))
(p2 = ggplot(data=iris) + geom_point(aes(x=Petal.Length, y=Petal.Width, color=Species)))

p = list(p1, p2)
(p3 = shareLegend(p, ncol = 2L, top = 'top_text', left = 'left_text'))
shareLegend(p, ncol = 1L, top = 'top_text', left = 'left_text', position = 'left')
shareLegend(list(p3, p3))
tryCatch(shareLegend(list(p3, p3), position = 'left'), error = identity)

## \CRANpkg{patchwork} is promising, but I dont know how to do what I want, yet
# library(patchwork)
# (tmp = p1 + p2 + plot_layout(guides = 'collect', nrow = 2, widths = 1)) # great
# tmp + labs(title = 'aa') # not what I want!!
}
\references{
\url{https://github.com/hadley/ggplot2/wiki/Share-a-legend-between-two-ggplot2-graphs}
}
\seealso{
Function \code{lemon::grid_arrange_shared_legend} also fetches
(via \code{lemon::g_legend})
the legend from 1st element of \code{grobs}.
This function does not have arguments \code{top}, \code{botom}, \code{left}, \code{right}.

Read more about Function \code{ggpubr::ggarrange}.

Read \url{https://plotly.com/r/subplots/}
on how to stack \code{'plotly'} objects!
}
