% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths_function.R
\name{split_list_}
\alias{split_list_}
\title{Re-organize a \link[base]{list} of Arguments}
\usage{
split_list_(x)
}
\arguments{
\item{x}{\link[base]{list} of \link[base]{vector}s}
}
\value{
Function \code{\link[=split_list_]{split_list_()}} returns a \link[base]{list}.
}
\description{
..
}
\examples{
list(mean = 1:2) |> split_list_() |> getval_()
list(mean = c(exp(1), pi), sd = 1:2) |> split_list_() |> getval_()
list(df1 = 1, df2 = c(2, 3), ncp = c(.1, .2)) |> split_list_() |> getval_()
}
\keyword{internal}
