% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoms_function.R
\name{getval_}
\alias{getval_}
\title{Values of Argument List}
\usage{
getval_(x, use_unicode = getOption("use_unicode"))
}
\arguments{
\item{x}{..}

\item{use_unicode}{\link[base]{logical} scalar}
}
\value{
Function \code{\link[=getval_]{getval_()}} returns a \link[base]{character} \link[base]{vector}.
}
\description{
..
}
\keyword{internal}
