% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoms_function.R
\name{getval_}
\alias{getval_}
\title{Values of Argument List}
\usage{
getval_(x)
}
\arguments{
\item{x}{..}
}
\value{
Function \link{getval_} returns a \link[base]{character} \link[base]{vector}.
}
\description{
..
}
