% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths_function.R
\name{paths_function}
\alias{paths_function}
\title{Multiple Function Geoms}
\usage{
paths_function(
  fun,
  dots = list(),
  args = .mapply(dots, FUN = list, MoreArgs = NULL),
  label,
  hjust = 0.5,
  parse = is.expression(label),
  aes_ = "colour",
  ...
)
}
\arguments{
\item{fun}{a continuous \link[base]{function}, see usage in function \link[ggplot2]{stat_function}}

\item{dots}{user-friendly \link[base]{list} of parameters,
e.g., \code{list(mean=0, sd=1:2)} for two normal distributions
\eqn{N(0,1)} and \eqn{N(0,2)}}

\item{args}{\link[base]{list} of parameters in a programmer-friendly manner,
e.g., \code{list(list(mean=0, sd=1), list(mean=0, sd=2))}
for two normal distributions
\eqn{N(0,1)} and \eqn{N(0,2)}}

\item{label}{\link[base]{character} \link[base]{vector},
user-friendly text description of the parameters}

\item{hjust}{\link[base]{double} scalar or \link[base]{vector}}

\item{parse}{\link[base]{logical} scalar}

\item{aes_}{\link[base]{character} scalar or \link[base]{vector},
\link[ggplot2]{aes}thetic mapping(s) to be activated.
Default \code{'colour'}, i.e., parameters are discriminated using different colors.
Another good choice is \code{c('colour', 'linetype')}, i.e., parameters are discriminated using different colors and line-types.}

\item{...}{argument of function \link[ggplot2]{stat_function}.
Also see parameter \code{params} of function \link[ggplot2]{layer}}
}
\value{
Function \code{\link[=paths_function]{paths_function()}} returns a \link[base]{list} of
\link[ggplot2]{layer}s.
}
\description{
Multiple function \emph{\strong{geom}}etric objects
(\link[ggplot2]{geom_function}s)
with different parameters.
}
\examples{
ggplot() + 
 paths_function(
  fun = dnorm, dots = list(mean = 1:0), hjust = .4,
  label = latex2exp::TeX(c('$mu_1=1$', '$mu_0=0$')) # bug; prefer unicode
 ) + 
 xlim(-3, 4)

}
\keyword{internal}
