% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoms_function.R
\name{split_list_}
\alias{split_list_}
\title{Re-organize a \link[base]{list} of Arguments}
\usage{
split_list_(x)
}
\arguments{
\item{x}{\link[base]{list} of \link[base]{vector}s}
}
\value{
Function \link{split_list_} returns a \link[base]{list}.
}
\description{
..
}
\examples{
getval_(split_list_(list(mean = 1:2)))
getval_(split_list_(list(mean = c(exp(1), pi), sd = 1:2)))
getval_(split_list_(list(df1 = 1, df2 = c(2, 3), ncp = c(.1, .2))))

}
